+++ examples/snake/src/logic/getInitialLine.ts
@@ -1,0 +1,37 @@
+import { Section, Point } from "./types.ts"
+import { getRandomInt, getRandomIntBetween } from "../lib/helpers.ts"
+import { boardSize } from "./logicConfig.ts"
+
+export function getInitialLine(): Section[] {
+  const startPoint = {
+    x: getRandomInt(boardSize.width),
+    y: getRandomInt(boardSize.height),
+  }
+
+  const boardCenterX = boardSize.width / 2
+  const boardCenterY = boardSize.height / 2
+
+  function getAngleLimits(startPoint: Point): [number, number] {
+    return startPoint.x < boardCenterX && startPoint.y < boardCenterY
+      ? [0, 90]
+      : startPoint.x > boardCenterX && startPoint.y < boardCenterY
+      ? [90, 180]
+      : startPoint.x < boardCenterX && startPoint.y > boardCenterY
+      ? [270, 360]
+      : startPoint.x > boardCenterX && startPoint.y > boardCenterY
+      ? [180, 270]
+      : [0, 360]
+  }
+
+  const angle = getRandomIntBetween(...getAngleLimits(startPoint))
+
+  return [
+    {
+      start: startPoint,
+      end: startPoint,
+      turning: "none",
+      endAngle: angle,
+      gap: false,
+    },
+  ]
+}
